% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare_helpers.R
\name{change_infoshare_dates}
\alias{change_infoshare_dates}
\title{Convert infoshare dates to YYYY-MM format}
\usage{
change_infoshare_dates(Date)
}
\arguments{
\item{Date}{The name of the column or object containing Infoshare-formatted dates}
}
\value{
A character vector with dates in YYYY-MM format
}
\description{
Infoshare files often contain dates ending in 'MXX' or 'QX'. This function converts these dates into YYYY-MM format.
Note: the output is not in a true date format as the underlying data may not refer to a specific date, but is easier to convert into dates as desired.
}
\examples{
\dontrun{
data(electronic_card_transactions)
electronic_card_transactionst$Dates <- change_infoshare_dates(electronic_card_transactions$Dates)
}
}
