% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electronic_card_transactions}
\alias{electronic_card_transactions}
\title{Monthly electronic card transaction values by industry group}
\format{
A data frame with 292 rows and 25 columns, in the default (wide) format for an Infoshare download. Variables are by default unlabelled.
}
\source{
Information: \url{https://datainfoplus.stats.govt.nz/item/nz.govt.stats/3d0dfc89-6bf5-4677-8cef-3befc36f5b90/45?_ga=2.17932670.636779380.1656904602-717270731.1599774632} Table: \url{https://infoshare.stats.govt.nz/} > Economic indicators > Electronic Card Transactions (ANZSIC06) - ECT > Values - Electronic card transactions A/S/T by industry group (Monthly)
}
\usage{
electronic_card_transactions
}
\description{
A dataset containing monthly electronic card transactions in New Zealand by industry group.
}
\keyword{datasets}
