% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_gdp_by_industry}
\alias{regional_gdp_by_industry}
\title{Gross domestic product by region and industry group}
\format{
A data frame with 51 rows and 469 columns, in the default (wide) format for an Infoshare download. Variables are by default unlabelled.
}
\source{
Information: \url{https://datainfoplus.stats.govt.nz/Item/example.org/3fd75697-5f34-4b49-aace-f961203ec6d2?_ga=2.6635893.636779380.1656904602-717270731.1599774632#/example.org/142b499e-c60b-4977-8236-355fe270c52c/2} Table: \url{https://infoshare.stats.govt.nz/} > Economic indicators > Regional Gross Domestic Product - RNA > Values - Gross domestic product, by region and industry (Annual-Mar)
}
\usage{
regional_gdp_by_industry
}
\description{
A dataset containing annual gross domestic product by region of New Zealand and industry group.
}
\keyword{datasets}
