% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare_helpers.R
\name{remove_infoshare_footnotes}
\alias{remove_infoshare_footnotes}
\title{Remove Infoshare footnotes}
\usage{
remove_infoshare_footnotes(df)
}
\arguments{
\item{df}{A dataframe, or an object coercible to a dataframe, with annoying Infoshare footnotes you would like to remove}
}
\value{
A dataframe without the Infoshare footnotes.
}
\description{
Remove Infoshare footnotes
}
\examples{
\dontrun{
data(electronic_card_transactions)
tail(electronic_card_transactions)
electronic_card_transactions <- remove_infoshare_footnotes(electronic_card_transactions)
}
}
