% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoshare_helpers.R
\name{wrangle_infoshare}
\alias{wrangle_infoshare}
\title{Tidy up data from Infoshare into a usable dataframe}
\usage{
wrangle_infoshare(
  df,
  dimension_names,
  measure_names = NA,
  old_measure_names = NA,
  coerce_to_numeric = TRUE,
  handle_dates = FALSE,
  long = TRUE,
  measures_in_row = NA
)
}
\arguments{
\item{df}{Required: a data frame, or object coercible to data frame, downloaded from infoshare using its CSV download function. The table downloaded must be in the default format provided by Infoshare and must not contain flags.}

\item{dimension_names}{A character vector containing the desired names of dimensions (qualitative / categorical variables).}

\item{measure_names}{A character vector containing the desired names of measures (quantitative / numerical variables). If unspecified, measure names will be used as they appear in the original table. If no measure names appear in the original table,}

\item{old_measure_names}{A character vector of measure names, as they appear in the original table. Used only to ensure measure_names replace old_measure_names in the appropriate order.}

\item{coerce_to_numeric}{Logical: coerce data values to numeric? Converts all suppressed or unavailable cells to NA. Defaults to TRUE.}

\item{handle_dates}{Logical: call handle_infoshare_dates? Defaults to FALSE.}

\item{long}{Logical: if TRUE, measure names will appear in a column named Measure and values will appear in a column named Value. If FALSE, each measure will appear in an individual column. Defaults to TRUE.}

\item{measures_in_row}{A single numeric value giving the table row in which measure names appear. Defaults to the row after the named dimensions, if there is a row containing measure names at all.}
}
\value{
A data frame with a column named Date, a column for each dimension, and either columns Measure and Value or one column per measure.
}
\description{
Tidy up data from Infoshare into a usable dataframe
}
\examples{
\dontrun{
data(regional_gdp_by_industry)
head(regional_gdp_by_industry)
wrangle_infoshare(
                  regional_gdp_by_industry,
                  dimension_names = c("Region", "Industry"),
                  measure_names = "GDP",
                  measures_in_row = 1,
                  long = FALSE
                  )
}

}
